create delete page and edit bootstrap file

+++ user_report.php
@@ -7,7 +7,7 @@
 <html>
 <head>
     <title>Crud Operation</title>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.1/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-+0n0xVW2eSR5OomGNYDnhzAbDsOXxcvSN1TPprVMTNDbiYZCxYbOOl7+AMvyTG2x" crossorigin="anonymous">
+    <link href="css/bootstrap.min.css" rel="stylesheet">
 </head>
 <body>
 <div  class="container-fluid">
@@ -19,7 +19,7 @@
                    <div class="card">
                        <div class="card-header">User Report</div>
                        <div CLASS="card-body">
-                        <table class="table" style="font-size: 12px;">
+                        <table class="table table-responsive-md" style="font-size: 12px;">
                             <tr>
                                 <th>SN</th>
                                 <th>Name</th>
@@ -41,7 +41,7 @@
                                 <td><?php echo $row["Email"]; ?></td>
                                 <td><?php echo $row["Address"]; ?></td>
                                 <td><a href="edit_user.php?uid=<?php echo $row["userID"]; ?>" class="btn btn-sm btn-Primary">Edit</a></td>
-                                <td><a href="delete_user.php" class="btn btn-sm btn-danger">Delete</a></td>
+                                <td><a href="delete_user.php?uid=<?php echo $row["userID"]; ?>" class="btn btn-sm btn-danger" onclick="return confirm('Are you sure you want to delete this record?')" >Delete</a></td>
                             </tr>
                         <?php } ?>
                             
@@ -68,8 +68,8 @@
 </div>
 
 
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.1/dist/js/bootstrap.bundle.min.js" integrity="sha384-gtEjrD/SeCtmISkJkNUaaKMoLD0//ElJ19smozuHV6z3Iehds+3Ulb9Bn9Plx0x4" crossorigin="anonymous"></script>
-<script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.9.2/dist/umd/popper.min.js" integrity="sha384-IQsoLXl5PILFhosVNubq5LC7Qb9DXgDA9i+tQ8Zj3iwWAwPtgFTxbJ8NT4GN1R8p" crossorigin="anonymous"></script>
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.1/dist/js/bootstrap.min.js" integrity="sha384-Atwg2Pkwv9vp0ygtn1JAojH0nYbwNJLPhwyoVbhoPwBhjQPR5VtM2+xf0Uwh9KtT" crossorigin="anonymous"></script>
+<script src="js/bootstrap.bundle.min.js"></script>
+<script src="js/popper.min.js"></script>
+<script src="js/bootstrap.min.js" ></script>
 </body>
 </html>
